refactor: move modal module up in scope
