+++ src/object-layer/dm-token-marker.js
@@ -5,7 +5,7 @@ import { TokenMarker } from "./token-marker";
 import { Input } from "../input";
 import * as Button from "../button";
 import * as Icon from "../feather-icons";
-import { Modal } from "../dm-area/modal";
+import { Modal } from "../modal";
 import { ToggleSwitch } from "../toggle-switch";
 import { useOvermind } from "../hooks/use-overmind";
